call hosted betterlink-no-interface.js script
Assuming that our core script is updated continuously, maintaining
frequent pushes to this repository will be difficult.

Using git subtrees is an unlikely solution, requiring too much git
expertise to use in a distributed environment without scripting. Pushing
commits to this repo would flood the commit log with noise, masking the
valuable commits made here.

It is unlikely that changes to the user interface will rely on specific
builds of the core script, so removing from source control seems like a
sensible option.