update filenames for clearer 'submissions' roles
Paired with a refactoring in the core Betterlink library, this update is
the beginning of an effort to give more control over the user-facing
look and feel to this library.

Additionally, 'highlighting' is specific to an implementation. The role
of submissions.viewer.js is to visualize a requested submission on the
page. That may include highlighting the content, but it's not required.
Similarly, submissions.interface.js may highlight content on the page to
confirm selection submission.