spike of in-place interface to submit selections
The basic interface gets out of the way of the user: provide 'bookends'
that enclose the range that could be submitted, but otherwise don't show
anything. On hover, identify that you can click on the selected content
to submit a new link.

Open issues:
- how to select content that is currently decorated
- use a different solution for :hover (because the hover only applies to
individual elements for a multi-piece range)
- submit the previous selection instead of relying on currentSelection()
- don't remove the selection after applying the 'prospectiveSubmission'
decoration
- create a loading animation
- allow Submissions.Result to cancel the loading animation
- choose a CSS style for the bookends