store the prospective submission for later send
When the user makes a prospective submission, store the result as an
array of ranges that can be submitted later. As part of this: use an
updated API that passes ranges around, instead of selection objects.

To handle multiple highlighters and multiple concurrent submissions
(possible with this UI approach), use an object-oriented approach so
that a Highlighter object maintains its state and cleans itself up when
no longer needed on the DOM.