add feature check for the CSS 'inherit' value
When supporting drag-and-drop, we will want to change the markup of our
'prospectiveSubmission' elements to be <a> anchors instead of spans.
This gives us wider browser support.

To create a clean UX, we will want these anchor links to blend into the
surrounding content. Most sites alter the default layout of anchors, so
we will use the 'inherit' value on the desired CSS attributes.

Using this check, we can fall back to an alternative approach if needed.