apply !important declaration on display:inline
As implemented, our reset CSS will never beat a specificity above
0,1,0,0 (ex: #my-div a). That's because we're styling multiple elements
and must use a class name, instead of an ID, to identify our elements.

If the element surrounding our highlights defines an alternative
'display' attribute, we'll break the formatting of the page upon
highlight. (ex: the 'Hot Network Questions' along the right side of
http://stackoverflow.com/)

That's why we specifically define our anchor CSS as display:inline.
However, we need to apply an !important declaration to override anchors
that are updated via ID attribution.