always enclose prospective submissions in new <a>
When using anchor tags to enclose a prospective submission, we should
never attempt to apply our class on existing elements. Instead, we
should always create a new wrapper anchor element.

The reason is so that we can treat these anchors like <span>s. We can
safely use 'inherit' for the properties we care about because we're
guaranteed for our new anchors to be children of the original source
elements.