remove the user selection during new submission
When the user clicks on a <span>, the default behavior is for the
browser to remove the current selection. This doesn't happen when a user
clicks on an anchor instead.

Because we're removing and adding highlights to the DOM during a
submission, the user selection can easily get messed up. The visual
noise is distracting. Instead, we should immediately remove the user
selection once they start a submission.