fire final dragend if element is removed from DOM
On submission we remove the anchor elements that enclose our selected
content. The drag events (that clients listen to) are subscribed onto
these elements.

Because the drop event fires before dragend, and we may execute a
submission during drop, the dragend event will not get called.

By specifically checking if any item is currently being dragged and if
it is about to be removed, then we can alert clients that the drag event
really *is* over.