store and clean up references to watched elements
Previously when elements were removed from the DOM, we would maintain
event listeners on these elements. It's much better to remove the
listeners.

In order to remove the listeners, we need to store references to the
elements being watched (which could maintain references long after the
elements are removed). This is fine as long as we alert the module that
the drag handlers need to be removed.

We can also now subscribe a turnOff() function to run if the full
application is being detached.