Revert "add Switchboard module to handle share actions"
This reverts commit 94d958a8f06f7dffe2d91a167e4c0149c6b45407.

The switchboard was an attempt to bypass browsers' popup restrictions.
However, this is ultimately going to fail.

By creating the switchboard, we create the popup on user action (which
the browsers will always allow). And then once the user decides *how*
they want to share, we'd redirect the window to take the appropriate
action.

The problem is that as part of the 'drop' action, the main window takes
focus. And as a continued part of preventing against popups, browsers
restrict our ability to target a particular window to give it focus.

This means that although we would always create the right action, the
user would never see the window below the taskbar. This is a bad user
experience.

Our alternatives:
- iframe the content into the page
** Not actually an option. Most services restrict iframes.
- Tell the user to find the hidden window
** Also not really an option
- Stick with the popups, but ask the user to enable popups
** Not an option. They'd have to enable popups for every page visited.
- Change the UX to require a click when the final share action happens
** This is what we'll do