fix state transition for lastSubmission
LastSubmission state machine updated to handle multiple asynchronous
events.

The submit() event always interrupts what we're doing to store our
latest submission. All future processing is skipped if the response
doesn't match this submission. So if we are in a terminal state (fail or
success) and a new response comes in (before we ever kicked back to the
submit state), then we can assume the response is out of order from a
prior submission.