trust any SVG location provided by build variable
Previously, we were checking if the build variable was updated to any
external resource that contained a period (like example.org). Instead,
we should trust any input that is provided, as long as the variable is
updated.

So the method for checking if we're running locally or in production is
whether the text of the build variable has changed.