refine state machine transitions during error state
If a submission failed but no response was received from the server (ex:
JSONP 404 for IE), then the state machine would error when trying to
make any subsequent submission. Instead, the state machine definition is
updated to support re-submissions in that situation.

Additionally, the UI was only showing the 'loading...' text any time the
LastSuccessful state machine transitioned into the submitted state. This
had the side effect of not showing the 'loading...' text when a second
submission was re-tried. The proper fix is to watch *any* submission
from the LastSubmission state machine, because this triggers the proper
subscriptions.