remove display of the link from Link Viewer
The content of the URL itself isn't useful to the user, as long as they
have clear options to access and share the URL.

This removes the display of the link, dedicates the text div to just the
selected content, and creates a dedicated div for an error message.

However, we're going to have an issue with the styles until we've
addressed an issue with the div-reset class taking precedence. This is
because we are no longer referring to these elements by id, but by
class.