refactor betterlink-reset CSS definitions
Moves the CSS reset definitions into a separate module (allowing us to
easily modify as needed) and defines the styles relative to all divs
underneath the root Action Drawer.

This has the side effect of requiring all additional selectors we define
to include the Action Drawer in order to beat the specificity of the
reset styles.

The pros: we no longer have to explicitly add a 'reset' class on
elements
The cons: we need to include the drawerSelector on all future
definitions (which necessitates a standard formatting of CSS styles to
prevent mistakes)