don't overlap header color prop with drawer border
When the drawer's border is normally applied, it overlaps the color prop
along the header. When the drawer is on top of a page with a dark
background, this is noticeable and distracting.

By applying the border as a pseudo element, we can overlay the border so
it starts underneath the header.