fix inheritance for div font color
Because of inheritance and specificity rules, a child div was not
inheriting the font color from a parent element. Instead, the color
specified in the general div reset was being applied.

Specifically stating that divs should inherit the font color fixes the
issue.

Also: removes the background redefinition for dropzones. No longer
necessary.