display up to three lines of selected text
Applies a CSS-only solution to clamp text to three lines, applying an
ellipsis for the overflow content. The ellipsis is actually applied on
top of the text content, and a gradient is applied to hide the text that
is covered.

A webkit-only approach could have achieved a similar effect with
-webkit-line-clamp.

Cons:
This solution requires a lot of math to be kept in sync. We also had to
adjust the font to a specific pixel-size to work across browsers.

Also, the ::before pseudo element maintains a constant height, which
means the action elements will be a long distance away for short
selections.