add function to explicitly provide image directory
This is needed when locally testing betterlink-interface on external
sites.

There needs to be a way for the local Chrome Extension to inform
betterlink_user_interface where the files are served from. This isn't an
issue when using the provided example_site.html because it's being
served from the same location.