conditionally displace body when sliding in drawer
The original approach to making room for the Action Drawer was to
displace the body of the webpage to the left, so that no content would
sit on top of the drawer. However, due to a WebKit bug
(https://code.google.com/p/chromium/issues/detail?id=445641) this
approach prematurely terminates drag events.

An alternative approach is to slide the drawer out, but apply a z-index
value so that the drawer will always sit on top of whatever content is
on the page.

Because client webpages may be using any values for z-index, we'll just
apply the maximum value.