use Anchor Reset as a base for betterlink-selected
Use the Anchor Reset CSS styles as a base in order to have the anchors
continue to blend into their surrounding text. This avoids line-height
or other changes that would cause the layout of the page to change.

inplace-anchor-highlighter uses a similar approach but is able to
cleanly define the CSS updates inline with the classnames. This is
because we use a separate class that is applied (via callback) on hover.
In order to use a single class, we have to merge our changes somehow
with the anchorResetCss. Here we use regexs to replace the content.