expand style initialization for the tooltip
Because our tooltip may be nested within the Action Drawer or outside of
it (within the rest of the page), we need to define the styles twice.
Once normally, and once scoped underneath the drawer in order to beat
the specificity of the drawer reset css.

This also adds a z-index to the tooltip for situations when it's outside
the drawer.