remove 'overflow: hidden' on the drawer
'overflow: hidden;' was defined in the layout originally from
http://blog.stevensanderson.com/2011/10/05/full-height-app-layouts-a-css-trick-to-make-it-easier/

However - it doesn't seem like this is necessary with the current
configuration. Likely because of the 'width:100%' definition for
betterlink-row.

Removing the overflow is going to be a necessary step for the tooltip,
allowing it to be positioned outside the drawer while being semantically
nested within.