add function to append tooltip to drawer
This creates a pair of functions: one that appends the tooltip to the
DOM as part of the drawer, and another for outside the drawer (the
positioning is different for each).

This will assume that we only have a single tooltip and that a client
will determine the content of the tooltip.