only execute FTE if user hasn't seen it before
We were previously checking to make sure we wouldn't load the FTE
multiple times while on the same page. However, this doesn't situations
where the FTE is loaded on multiple pages (or the user just refreshes
the page).

Instead, alert the API server that the user has seen the FTE and query
the server before loading the resources. This will allow us to better
guard against showing the FTE to the same user multiple times.