add section to test copying/pasting text
There's a current problem that highlighting text and copying it will
copy the HTML <a> elements as well. This is a problem if the target is
an HTML text area (like an email compose window).

This creates a content-editable <div> that can be used for testing. If
an anchor tag is copied into the text area, a red rectangle will appear
underneath it. (we can't style the anchors themselves because the
content-editable <div> is too smart and applies inline styles to
compensate)