supply nonce to style attributes if provided
A nonce is used as a one-time, randomly-generated string. In this
instance it is used to indicate style elements that are 'approved' by a
site's Content Security Policy.
(http://www.w3.org/TR/CSP2/#script-src-nonce-usage)

If a nonce has been provided to Betterlink, then apply it to all style
elements we create.