switch opacity for padding in .betterlink-highlight
Opacity was added in cd83d1f to resolve issues when the highlights
overlapped each other, blocking text. The opacity created separate
stacking contexts which resolved the problem.

However, the root of the problem was that some frameworks (like
Bootstrap) apply a slight padding (0.2em) to <mark> elements. Removing
this padding resolves the overlapping issue (and other minor layout
changes).

Also:
Removing the opacity solves a potential issue when using the FTE. If the
selected text that triggers the FTE is within of a .betterlink-highlight
element, then the tooltip will get hidden behind the overlay, rendering
it unclickable. This is because the <mark> element's opacity creates a
separate stacking context that prevents its children (the tooltip) from
ever stacking above other elements.

This would still be a problem for selections that are part of other
elements with an opacity, but that has a lower probability.