modify FTE demo link to indicate it's from FTE
When a user is demoing the new link they created by clicking "Follow
Your Link" on the FTE tooltip, modify the URL slightly.

This modification will inform the Betterlink script on the subsequent
page load that the user is coming from the FTE. This can serve as an
indicator to provide more guidance to the user.