use a cookie to indicate coming from the FTE
This reverts 8530e0e and replaces it with an implementation that uses a
cookie to indicate referral from the FTE (instead of a URL param).

The major concern is that by modifying the URL, it would be non-trivial
to prevent the added parameters from staying within the URL if the user
were to make a submission from that page.

By using a cookie, we don't run into that issue because we can safely
remove the cookie from within the same module. Additionally, we can set
an expiration on the cookie so that it'll remove itself naturally as
well.

The downside:
Not all users support cookies.