only run copy protection if active highlighters
This was specifically a problem if trying to copy the submission URL
from the 'copy link' element. We had highlighted content, but because it
was the text within a textbox input, there was no highlighter markup or
DOM selection. This was accidentally canceling the selection and wasn't
replacing it, preventing the copy.