disable Facebook element
If a user shares a link from Betterlink via Facebook, the link will be
collapsed into the canonical form of that URL. This means that our
unique URL parameter will disappear from the link that was shared.

So if someone attempted to share the URL this way, the result will not
be what they're expecting. So the better option is to visually disable
the feature.