adjust z-index so tooltip always above content
In the event that the FTE tooltip is applied to the page, but below the
content its pointing to (ex: the content is at the very top of the
page), then the tooltip would end up being layered between the overlay
and the content.

We always want the tooltip to rest on top, so we need to adjust the
z-indexes to create more differentiation.